+++ core/modules/modBankImport.class.php
@@ -15,7 +15,7 @@ class modBankImport extends DolibarrModules
         global $langs, $conf;
         $this->db = $db;
 
-        $this->version = BankImportHelper::getEnv('VERSION', '0.0.10');
+        $this->version = BankImportHelper::getEnv('VERSION', '0.0.11');
 
         // Unique ID (custom modules > 100000)
         $this->numero = 104001;
@@ -47,9 +47,8 @@ class modBankImport extends DolibarrModules
         $r = 0;
         $this->rights[$r][0] = $this->numero + $r;
         $this->rights[$r][1] = 'Bankauszüge importieren';
-        $this->rights[$r][2] = 'w';
-        $this->rights[$r][3] = 0;
-        $this->rights[$r][4] = 'modifier';
+        $this->rights[$r][3] = 1;
+        $this->rights[$r][4] = 'import';
         $r++;
 
         // --- Menu definition ---
@@ -64,7 +63,7 @@ class modBankImport extends DolibarrModules
             'langs'     => 'bankimport@bankimport',
             'position'  => 100,
             'enabled'   => '1',
-            'perms'     => '$user->hasRight("banque", "modifier")',
+            'perms'     => '1',
             'target'    => '',
             'user'      => 0
         );